Adding PDF for session-2

+++ batch_four/session-2/addition.py
@@ -3,8 +3,8 @@
 '''Conversion of strings to int before addition'''
 
 xString = input("Enter a number: ")
-x = int(xString)
+x = float(xString)
 yString = input("Enter a second number: ")
-y = int(yString)
+y = float(yString)
 print('The sum of ', x, ' and ', y, ' is ', x+y, '.', sep='|')
